+++ servicetalk-concurrent-api/src/main/java/io/servicetalk/concurrent/api/ContextCaptureProvider.java
@@ -29,5 +29,12 @@ interface ContextCaptureProvider {
      */
     CapturedContext captureContext();
 
+    /**
+     * Save a copy of the existing context in preparation for an asynchronous thread jump.
+     *
+     * Note that this can do more than just package up the ServiceTalk {@link AsyncContext} and could be enhanced or
+     * wrapped to bundle up additional contexts such as the OpenTelemetry or grpc contexts.
+     * @return the copied saved context state that may be restored later.
+     */
     CapturedContext captureContextCopy();
 }
