+++ servicetalk-concurrent-api/src/main/java/io/servicetalk/concurrent/api/PublisherSetContextOnSubscribe.java
@@ -40,7 +40,7 @@ final class PublisherSetContextOnSubscribe<T> extends AbstractNoHandleSubscribeP
             @Override
             public Scope restoreContext() {
                 Scope outer = parentContext.restoreContext();
-                Scope inner = ContextMapThreadLocal.attachContext(context);
+                Scope inner = AsyncContext.provider().attachContext(context);
                 return () -> {
                     inner.close();
                     outer.close();
