+++ servicetalk-concurrent-api/src/main/java/io/servicetalk/concurrent/api/SingleSetContextOnSubscribe.java
@@ -40,7 +40,7 @@ final class SingleSetContextOnSubscribe<T> extends AbstractNoHandleSubscribeSing
             @Override
             public Scope restoreContext() {
                 Scope outer = parentContext.restoreContext();
-                Scope inner = ContextMapThreadLocal.attachContext(context);
+                Scope inner = AsyncContext.provider().attachContext(context);
                 return () -> {
                     inner.close();
                     outer.close();
